first commit setting choices slides
