+++ homebrewing/src/components/Navbar/Choices/Choices.css
@@ -17,7 +17,7 @@
 }
 
 .canvas {
-  width: 20% !important;
+  width: 25% !important;
 }
 
 .offcanvas {
@@ -171,4 +171,4 @@
   .canvasBody .nav .links {
     font-size: 30px !important;
   }
-}/*# sourceMappingURL=Nav.css.map */
+}/*# sourceMappingURL=Choices.css.map */
