+++ homebrewing/src/components/Navbar/Choices/Choices.scss
@@ -18,7 +18,7 @@
 }
 
 .canvas {
-    width: 20%!important;
+    width: 25%!important;
 }
 
 .offcanvas {
